+++ resources/js/Pages/Auth/ForgotPassword.tsx
@@ -1,5 +1,5 @@
 import GuestLayout from "@/Layouts/GuestLayout";
-import InputError from "@/Components/input-error";
+import InputError from "@/Components/InputError";
 import PrimaryButton from "@/Components/PrimaryButton";
 import TextInput from "@/Components/TextInput";
 import { Head, useForm } from "@inertiajs/react";
