+++ resources/js/Pages/Profile/Partials/UpdateProfileInformationForm.tsx
@@ -1,4 +1,4 @@
-import InputError from "@/Components/input-error";
+import InputError from "@/Components/InputError";
 import InputLabel from "@/Components/InputLabel";
 import PrimaryButton from "@/Components/PrimaryButton";
 import TextInput from "@/Components/TextInput";
